requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.5
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-modeladmin-reorder==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.23.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0

